backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
email-validator
pypdf
sentence_transformers
faiss-cpu
openai
requests

